Refactor unpack strategies into separate files.
