+++ Library/Homebrew/test/unpack_strategy/bzip2_spec.rb
@@ -1,0 +1,8 @@
+require_relative "shared_examples"
+
+describe UnpackStrategy::Bzip2 do
+  let(:path) { TEST_FIXTURE_DIR/"cask/container.bz2" }
+
+  include_examples "UnpackStrategy::detect"
+  include_examples "#extract", children: ["container"]
+end
