+++ Library/Homebrew/test/unpack_strategy/cvs_spec.rb
@@ -1,0 +1,14 @@
+require_relative "shared_examples"
+
+describe UnpackStrategy::Cvs do
+  let(:repo) {
+    mktmpdir.tap do |repo|
+      FileUtils.touch repo/"test"
+      (repo/"CVS").mkpath
+    end
+  }
+  let(:path) { repo }
+
+  include_examples "UnpackStrategy::detect"
+  include_examples "#extract", children: ["CVS", "test"]
+end
