+++ Library/Homebrew/test/unpack_strategy/jar_spec.rb
@@ -1,0 +1,8 @@
+require_relative "shared_examples"
+
+describe UnpackStrategy::Jar do
+  let(:path) { TEST_FIXTURE_DIR/"test.jar" }
+
+  include_examples "UnpackStrategy::detect"
+  include_examples "#extract", children: ["test.jar"]
+end
