+++ Library/Homebrew/test/unpack_strategy/lha_spec.rb
@@ -1,0 +1,7 @@
+require_relative "shared_examples"
+
+describe UnpackStrategy::Lha do
+  let(:path) { TEST_FIXTURE_DIR/"test.lha" }
+
+  include_examples "UnpackStrategy::detect"
+end
