+++ Library/Homebrew/test/unpack_strategy/lzip_spec.rb
@@ -1,0 +1,7 @@
+require_relative "shared_examples"
+
+describe UnpackStrategy::Lzip do
+  let(:path) { TEST_FIXTURE_DIR/"test.lz" }
+
+  include_examples "UnpackStrategy::detect"
+end
