+++ Library/Homebrew/test/unpack_strategy/p7zip_spec.rb
@@ -1,0 +1,7 @@
+require_relative "shared_examples"
+
+describe UnpackStrategy::P7Zip do
+  let(:path) { TEST_FIXTURE_DIR/"cask/container.7z" }
+
+  include_examples "UnpackStrategy::detect"
+end
