+++ Library/Homebrew/test/unpack_strategy/rar_spec.rb
@@ -1,0 +1,7 @@
+require_relative "shared_examples"
+
+describe UnpackStrategy::Rar do
+  let(:path) { TEST_FIXTURE_DIR/"cask/container.rar" }
+
+  include_examples "UnpackStrategy::detect"
+end
