+++ Library/Homebrew/test/unpack_strategy/xz_spec.rb
@@ -1,0 +1,7 @@
+require_relative "shared_examples"
+
+describe UnpackStrategy::Xz do
+  let(:path) { TEST_FIXTURE_DIR/"cask/container.xz" }
+
+  include_examples "UnpackStrategy::detect"
+end
