+++ Library/Homebrew/unpack_strategy/bazaar.rb
@@ -1,0 +1,18 @@
+require_relative "directory"
+
+module UnpackStrategy
+  class Bazaar < Directory
+    def self.can_extract?(path:, magic_number:)
+      super && (path/".bzr").directory?
+    end
+
+    private
+
+    def extract_to_dir(unpack_dir, basename:, verbose:)
+      super
+
+      # The export command doesn't work on checkouts (see https://bugs.launchpad.net/bzr/+bug/897511).
+      FileUtils.rm_r unpack_dir/".bzr"
+    end
+  end
+end
