+++ Library/Homebrew/unpack_strategy/cvs.rb
@@ -1,0 +1,9 @@
+require_relative "directory"
+
+module UnpackStrategy
+  class Cvs < Directory
+    def self.can_extract?(path:, magic_number:)
+      super && (path/"CVS").directory?
+    end
+  end
+end
