+++ Library/Homebrew/unpack_strategy/git.rb
@@ -1,0 +1,9 @@
+require_relative "directory"
+
+module UnpackStrategy
+  class Git < Directory
+    def self.can_extract?(path:, magic_number:)
+      super && (path/".git").directory?
+    end
+  end
+end
