+++ Library/Homebrew/unpack_strategy/uncompressed.rb
@@ -1,0 +1,13 @@
+module UnpackStrategy
+  class Uncompressed
+    include UnpackStrategy
+
+    alias extract_nestedly extract
+
+    private
+
+    def extract_to_dir(unpack_dir, basename:, verbose:)
+      FileUtils.cp path, unpack_dir/basename, preserve: true, verbose: verbose
+    end
+  end
+end
